Move CLI helpers out of ds
