+++ src/clients/perf/perf_client.h
@@ -2,7 +2,7 @@
 // Licensed under the Apache 2.0 License.
 #pragma once
 
-#include "ds/cli_helper.h"
+#include "common/cli_helper.h"
 
 #include <CLI11/CLI11.hpp>
 #include <chrono>
