+++ src/host/run.cpp
@@ -16,12 +16,12 @@
 #include "ccf/pal/snp_ioctl.h"
 #include "ccf/service/node_info_network.h"
 #include "ccf/version.h"
+#include "common/cli_helper.h"
 #include "common/configuration.h"
 #include "common/enclave_interface_types.h"
 #include "config_schema.h"
 #include "configuration.h"
 #include "crypto/openssl/hash.h"
-#include "ds/cli_helper.h"
 #include "ds/files.h"
 #include "ds/internal_logger.h"
 #include "ds/non_blocking.h"
